requirements.txt
rasterio
gdal
pyyaml
click
py-dateutil
